chore(multi-site): Preliminary Site Manager Access control

+++ http/api_admin_cache_flush.go
@@ -16,6 +16,10 @@ func (a *action) AdminCacheFlush(c echo.Context) error {
 		return resp
 	}
 
+	if !GetIsSuperAdmin(c) {
+		return RespError(c, "无权访问")
+	}
+
 	if p.FlushAll {
 		go func() {
 			model.CacheFlushAll()
